refactor(core): extract case analysis orchestrator from history module
